examples/aiohttp_jinja2/requirements.txt
aiohttp-jinja2
aiohttp-etag
